refactor: move poll and vote handling from ActiveRole to PassiveRole
Pure structural move of handlePoll, handleVote and isLogUpToDate into the
superclass, so that passive and promotable members can answer polls and votes
through the exact same code path in a follow-up. No behavior change:
PassiveRole.onPoll and onVote still reject with ILLEGAL_MEMBER_STATE.
